Generate `__init__.py` files for enum-only modules
The current Python code generator does not emit `__init__.py` files for
modules that only contain enumerations. This is problematic in cases
such as https://github.com/pulumi/pulumi/issues/16221, whereby
`pulumi-command` wishes to define an enumeration to be shared across
modules, in a module with no other types or resources. In such cases, an
SDK is generated that does not type check, since the enumeration types
are not exported appropriately. This commit addresses this by generating
`__init__.py` files in these cases as expected.