fix panic with --continue-on-error on delete after failed create
For deleting resources when `--continue-on-error` is given, we look up
a failed resources transitive dependencies in the dependency graph.
This is used to skip those resources from being destroyed to keep the
dependency chains correctly intact.

The errors for looking these dependencies up are recorded in the step
executor, whenever a resource execution fails.  This includes errors
from resources that are supposed to be created.

However when a resource fails to create that resource does not exist
in the dependency graph at all, so we cannot look it up, and we get a
panic because we try to access a nil pointer.

When deleting resources we do not need to care about resources that
failed earlier, so we can just clear the errored steps before going
ahead with the deletes.