update homebrew tap push to use `main` (#16262)
We're in the process of renaming the main branch in our repos to `main`
from `master`. I'm planning to do homebrew-tap next, and this `git push`
will need to be updated to keep that working.

This is meant to be merged right after the branch is renamed in that
repo.