fix panic with --continue-on-error on delete after failed create (#16261)
For deleting resources when `--continue-on-error` is given, we look up a
failed resources transitive dependencies in the dependency graph. This
is used to skip those resources from being destroyed to keep the
dependency chains correctly intact.

The errors for looking these dependencies up are recorded in the step
executor, whenever a resource execution fails. This includes errors from
resources that are supposed to be created.

However when a resource fails to create that resource does not exist in
the dependency graph at all, so we cannot look it up, and we get a panic
because we try to access a nil pointer.

When deleting resources we do not need to care about resources that
failed earlier, so we can just clear the errored steps before going
ahead with the deletes.

Fixes https://github.com/pulumi/pulumi/issues/16258

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->