Normalize plugin.Provider methods to (Context, Request) -> (Response, error)
Normalize methods on plugin.Provider to the form:

```go
Method(context.Context, MethodRequest) (MethodResponse, error)
```

This provides a more consistent and forwards compatible interface for each of our methods.

---

I'm motivated to work on this because the bridge maintains a copy of this interface:
`ProviderWithContext`. I'm willing to fix consumers of `plugin.Provider` in
`pulumi/pulumi`, but I wanted to make sure that we would be willing to merge this PR if I
get it green.