improve error handling when downloading policy packs
The HTTP client in Go only returns an error if there is a protocol
level failure, or an error caused by client policy (which we don't
use).  However the policy pack can fail to be downloaded, and return a
non 200 HTTP status.  This should still be considered an error.

Currently we would only detect this when trying to unpack the policy
pack.  Catching and returning this error earlier allows us to give
better information about the failure to the user.