cli: buffer output to system pager (#327)
Experimentally, this helps avoid problems with corrupted output when
piping to the system pager.

These changes also disable the `bubbletea`-based fallback pager,
instead falling back to no paging. The `bubbletea` pager just doesn't
seem to work properly, and we need a quick fix here. Hopefully we can
get a better root cause for all of this in the future.