cli: add env version retract (#330)
This command retracts a specific revision of an environment. A retracted
revision can no longer be read or opened. Retracting a revision also
updates any tags that point to the retracted revision to instead point
to a replacement revision. If no replacement is specified, the latest
non-retracted revision preceding the revision being retracted is used as
the replacement.

The revision pointed to by the `latest` tag may not be retracted. To
retract the latest revision of an environment, first update the
environment with a new definition.