only deny creation for default providers, not fetching it
When the `disable-default-providers` settings is set, we want to make
sure to not create any new default providers implicitly.  However
programmatic default providers are explicitly set, but they are using
the same `defaultProviders` implementation.  When we have already set
a default provider through the programmatic default provider
mechanism, we don't want to error out here.

It's safe to just move the loading of the default provider above the
denyCreation check, as we won't set an implicit default provider if
this setting is set, so we'll always hit the denyCreation code.