[go/program-gen] Fix union type type resolution in Go program generation (#16297)
# Description

Fixes https://github.com/pulumi/pulumi-azure-native/issues/1554 

### Context
The problem here is that when we compute `InputType: model.Type` in
`pcl.Resource`, we map the types of input properties of resources from
`schema.Type` into `model.Type`. When one of these properties is a
`schema.UnionType` (union of objects to be exact), we map that _as is_
to `model.UnionType` which trips up Go program-gen as it doesn't know
how to reduce the type to the actual _selected_ object type based on the
resource inputs.

### Resolution
The way to fix this is not in Go program-gen, instead we _reduce_ the
computed union types during the binding phase into the actual object
types based on the resource inputs so that all program generators only
work against explicit objects rather than having to deal with unions of
objects

### Example:

```pcl
resource "example" "azure-native:eventgrid:EventSubscription" {
    destination = {
        endpointType = "EventHub"
        resourceId = "example"
    }
    expirationTimeUtc = "example"
    scope = "example"
}
```

Before:
```go
pulumi.Run(func(ctx *pulumi.Context) error {
	_, err := eventgrid.NewEventSubscription(ctx, "example", &eventgrid.EventSubscriptionArgs{
		Destination: eventgrid.EventHubEventSubscriptionDestination{
			EndpointType: "EventHub",
			ResourceId:   "example",
		},
		ExpirationTimeUtc: pulumi.String("example"),
		Scope:             pulumi.String("example"),
	})
	if err != nil {
		return err
	}
	return nil
})
```

After:
```go
pulumi.Run(func(ctx *pulumi.Context) error {
	_, err := eventgrid.NewEventSubscription(ctx, "example", &eventgrid.EventSubscriptionArgs{
		Destination: &eventgrid.EventHubEventSubscriptionDestinationArgs{
			EndpointType: pulumi.String("EventHub"),
			ResourceId:   pulumi.String("example"),
		},
		ExpirationTimeUtc: pulumi.String("example"),
		Scope:             pulumi.String("example"),
	})
	if err != nil {
		return err
	}
	return nil
})
```

## Checklist

- [ ] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [ ] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->