wait for RPCs to finish before registering stack transforms
For transforms, we want to make sure all resources are registered
before registering them, and that no resource registration happens
while the transform is being registered in the engine.

The latter is already given, since `register_stack_transform` happens
synchronously, so no new RegisterResource calls can be started while
the transforms are registered.

For the former, we can make use of the event loop.  Since all of the
rpc calls are being put on the event loop, we can make sure to wait
for them to finish before re start registering the transforms.  Note
that we can't just wait for all tasks to finish because there might be
more than RPCs in flight.  However when the RPCs finished we can be
certain that the resources are registered in the engine, and
registering the stack transforms is now a valid operation.