Compute refresh diffs based on desired state
Presently, the behaviour of diffing during refresh steps is incomplete,
returning only an "output diff" that presents the changes in outputs.
This commit changes refresh steps so that:

* they compute a diff similar to the one that would be computed if a
  `preview` were run immediately after the refresh, which is more
  typically what users expect and want; and
* `IgnoreChanges` resource options are respected when performing the new
  desired-state diffs, so that property additions or changes reported by
  a refresh can be ignored.

In particular, `IgnoreChanges` can now be used to acknowledge that part
or all of a resource may change in the provider, but the user is OK with
this and doesn't want to be notified about it during a refresh.
Importantly, this means that the diff won't be reported, but also that
the changes won't be applied to state.

The implementation covers the following:

* A diff is computed using the inputs from the program and then
  inverting the result, since in the case of a refresh the diff is being
  driven by the provider side and not the program. This doesn't change
  what is stored back into the state, but it does produce a diff that is
  more aligned with the "true changes to the desired state".
* `IgnoreChanges` resource options are now stored in state, so that this
  information can be used in refresh operations that do not have access
  to/run the program.
* In the context of a refresh operation, `IgnoreChanges` applies to
  *both* input and output properties. This differs from the behaviour of
  a normal update operation, where `IgnoreChanges` only considers input
  properties.
* The special `"*"` value for `IgnoreChanges` can be used to ignore all
  properties. It _also_ ignores the case where the resource cannot be
  found in the provider, and instead keeps the resource intact in state
  with its existing input and output properties.

Because the program is not run for refresh operations, `IgnoreChanges`
options must be applied separately before a refresh takes place. This
can be accomplished using e.g. a `pulumi up` that applies the options
prior to a refresh. In future we could introduce a `pulumi state set
...` command or the like to modify `IgnoreChanges` resource options in
state directly, but this commit does not address this.

For use cases relying on the legacy refresh diff provider, the
`PULUMI_USE_LEGACY_REFRESH_DIFF` environment variable can be set, which
will disable desired-state diff computation. We only need to perform
checks in `RefreshStep.{ResultOp,Apply}`, since downstream code will
work correctly based on the presence or absence of a `DetailedDiff` in
the step.

Co-authored-by: Luke Hoban <lukehoban@gmail.com>