Update current nodejs version to 22.x (#16378)
# Description

The current nodejs version is 22.x
https://nodejs.org/en/about/previous-releases

This might help us with GH actions failing on the node download step.
The node-setup action will download LTS versions from
https://github.com/actions/node-versions/blob/main/versions-manifest.json,
and fallback to the official nodejs.org download if not found. Since we
had node set to 21.x before, it was always downloading from nodejs.org,
which seems to be fairly flakey.

This won't fix it once we test against 23.x though :/